refactor: [#18] 영화 관련된 파일을 폴더로 정리
