+++ Web 5주차/pages/MoviePages/NowPlayingPage.jsx
@@ -1,4 +1,4 @@
-import MoviesFetchComponent from '../components/MoviesFetchComponent';
+import MoviesFetchComponent from '../../components/MoviesFetchComponent';
 
 function NowPlayingPage() {
     const address = 'https://api.themoviedb.org/3/movie/now_playing?language=ko-KRx&page=1';
