+++ Web 5주차/pages/MoviePages/PopularPage.jsx
@@ -1,4 +1,4 @@
-import MoviesFetchComponent from '../components/MoviesFetchComponent';
+import MoviesFetchComponent from '../../components/MoviesFetchComponent';
 
 function PopularPage() {
     const address = 'https://api.themoviedb.org/3/movie/popular?language=ko-KR&page=1';
