+++ Web 5주차/pages/MoviePages/TopRatedPage.jsx
@@ -1,4 +1,4 @@
-import MoviesFetchComponent from '../components/MoviesFetchComponent';
+import MoviesFetchComponent from '../../components/MoviesFetchComponent';
 
 function TopRatedPage() {
     const address = 'https://api.themoviedb.org/3/movie/top_rated?language=ko-KR&page=1';
